Reorganize UI module
