--- src/ButtonHandler.ts
@@ -1,3 +1,0 @@
-import { ButtonInteraction } from 'discord.js';
-
-export type ButtonHandler = (interaction: ButtonInteraction) => void | Promise<void>;
