+++ src/index.ts
@@ -1,4 +1,3 @@
-export * from './ButtonHandler';
 export { default as Client } from './Client';
 export * from './Command';
 export * from './dispatch';
